Linting and refactoring
