+++ crackle.config.ts
@@ -3,5 +3,5 @@ import { defineConfig } from '@crackle/cli/config';
 export default defineConfig({
   package: {
     mode: 'bundle',
-  },
+  }
 });
