+++ package.json
@@ -36,24 +36,31 @@
     "dist"
   ],
   "scripts": {
-    "build": "pnpm crackle:package ; pnpm rename && pnpm web-ext:build && pnpm repack-for-chromium",
+    "build": "pnpm clean ; pnpm lint ; pnpm crackle:package ; pnpm rename && pnpm web-ext:build && pnpm repack-for-chromium",
     "crackle:package": "crackle package",
+    "clean": "rm -r ./chromium/ ./dist/ ./eventuate-chromium.zip ./web-ext-artifacts/",
     "rename": "bash -c 'mv ./dist/index.cjs ./dist/index.js'",
     "repack-for-chromium": "bash './script/repack-for-chromium'",
     "test": "jest",
+    "lint": "eslint .",
     "web-ext:build": "web-ext build -o",
     "web-ext:lint": "web-ext lint",
     "start": "web-ext run -v -u https://www.parkrun.com.au/coburg/results/latestresults/"
   },
   "devDependencies": {
     "@crackle/cli": "^0.15.5",
+    "@eslint/js": "^9.13.0",
+    "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.13",
     "@types/node": "^22.7.5",
+    "eslint": "^9.13.0",
+    "globals": "^15.11.0",
     "jest": "^29.7.0",
     "prettier": "^3.3.3",
     "ts-jest": "^29.2.5",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.6.2",
+    "typescript-eslint": "^8.10.0",
     "web-ext": "^8.3.0"
   },
   "packageManager": "pnpm@9.2.0",
