+++ src/types/Volunteer.ts
@@ -1,0 +1,7 @@
+export interface Volunteer {
+  name: string;
+  link: string;
+  athleteID: number;
+  vols?: string;
+  agegroup?: string;
+}
